+++ hdf/test/tusejpegfuncs.c
@@ -175,7 +175,7 @@ decomp_using_jpeglib(const char *filename,    /* file to read compressed data fr
 
     /* Allocate local buffer to hold read values until all reading is done
        before copying into caller's buffer */
-    local_buf = malloc(im_height * im_width * im_ncomps * sizeof(uint8));
+    local_buf = malloc((size_t)(im_height * im_width * im_ncomps) * sizeof(uint8));
     CHECK_ALLOC(local_buf, "local_buf", "decomp_using_jpeglib");
 
     /* Set up the JPEG error routines */
@@ -214,12 +214,12 @@ decomp_using_jpeglib(const char *filename,    /* file to read compressed data fr
         (void)jpeg_read_scanlines(&cinfo, buffer, 1);
 
         /* Saved read line to the local buffer */
-        memcpy(ptr, buffer[0], row_stride);
+        memcpy(ptr, buffer[0], (size_t)row_stride);
         ptr = ptr + row_stride;
     }
 
     /* Copying values from local buffer to caller's buffer after success */
-    memcpy(read_buffer, local_buf, im_height * im_width * im_ncomps);
+    memcpy(read_buffer, local_buf, (size_t)(im_height * im_width * im_ncomps));
     free(local_buf);
 
     /* Finish decompression */
