+++ hdf/util/he_cntrl.c
@@ -326,7 +326,7 @@ HEIdump(int32 length, int offset, char *format, int raw_flag)
 
         case 'i': {
             int32 *idata;
-            idata = (int32 *)malloc(length / 4 * sizeof(int32));
+            idata = (int32 *)malloc((size_t)(length / 4) * sizeof(int32));
 
             DFKconvert((void *)(data + offset), (void *)idata, DFNT_INT32 | raw_flag, length / 4, DFACC_READ,
                        0, 0);
@@ -344,7 +344,7 @@ HEIdump(int32 length, int offset, char *format, int raw_flag)
 
         case 'd': {
             uint32 *idata;
-            idata = (uint32 *)malloc(length / 4 * sizeof(int32));
+            idata = (uint32 *)malloc((size_t)(length / 4) * sizeof(int32));
 
             DFKconvert((void *)(data + offset), (void *)idata, DFNT_UINT32 | raw_flag, length / 4, DFACC_READ,
                        0, 0);
@@ -361,7 +361,7 @@ HEIdump(int32 length, int offset, char *format, int raw_flag)
         } break;
         case 'j': {
             int16 *sdata;
-            sdata = (int16 *)malloc(length / 2 * sizeof(int16));
+            sdata = (int16 *)malloc((size_t)(length / 2) * sizeof(int16));
             DFKconvert((void *)(data + offset), (void *)sdata, DFNT_INT16 | raw_flag, length / 2, DFACC_READ,
                        0, 0);
 
@@ -379,7 +379,7 @@ HEIdump(int32 length, int offset, char *format, int raw_flag)
 
         case 's': {
             uint16 *sdata;
-            sdata = (uint16 *)malloc(length / 2 * sizeof(uint16));
+            sdata = (uint16 *)malloc((size_t)(length / 2) * sizeof(uint16));
 
             DFKconvert((void *)(data + offset), (void *)sdata, DFNT_UINT16 | raw_flag, length / 2, DFACC_READ,
                        0, 0);
@@ -397,7 +397,7 @@ HEIdump(int32 length, int offset, char *format, int raw_flag)
 
         case 'b': {
             uint8 *bdata;
-            bdata = (uint8 *)malloc(length);
+            bdata = (uint8 *)malloc((size_t)length);
 
             DFKconvert((void *)(data + offset), (void *)bdata, DFNT_UINT8 | raw_flag, length, DFACC_READ, 0,
                        0);
@@ -418,7 +418,7 @@ HEIdump(int32 length, int offset, char *format, int raw_flag)
             size_t sizeint;
 
             sizeint = sizeof(int);
-            idata   = (int *)malloc((size_t)(length / 4 * sizeint));
+            idata   = (int *)malloc((size_t)(length / 4) * sizeint);
             DFKconvert((void *)(data + offset), (void *)idata, DFNT_NINT32 | raw_flag, length / 4, DFACC_READ,
                        0, 0);
             printf("%8d: ", offset);
@@ -438,7 +438,7 @@ HEIdump(int32 length, int offset, char *format, int raw_flag)
             size_t sizeint;
 
             sizeint = sizeof(int);
-            idata   = (int *)malloc((size_t)(length / 4 * sizeint));
+            idata   = (int *)malloc((size_t)(length / 4) * sizeint);
             DFKconvert((void *)(data + offset), (void *)idata, DFNT_NINT32 | raw_flag, length / 4, DFACC_READ,
                        0, 0);
             printf("%8d: ", offset);
@@ -472,7 +472,7 @@ HEIdump(int32 length, int offset, char *format, int raw_flag)
 
         case 'f': {
             float32 *fdata;
-            fdata = (float32 *)malloc(length / 4 * sizeof(float32));
+            fdata = (float32 *)malloc((size_t)(length / 4) * sizeof(float32));
 
             DFKconvert((void *)(data + offset), (void *)fdata, DFNT_FLOAT32 | raw_flag, length / 4,
                        DFACC_READ, 0, 0);
@@ -491,7 +491,7 @@ HEIdump(int32 length, int offset, char *format, int raw_flag)
 
         case 'e': {
             float64 *fdata;
-            fdata = (float64 *)malloc(length / 8 * sizeof(float64));
+            fdata = (float64 *)malloc((size_t)(length / 8) * sizeof(float64));
 
             DFKconvert((void *)(data + offset), (void *)fdata, DFNT_FLOAT64 | raw_flag, length / 8,
                        DFACC_READ, 0, 0);
@@ -614,7 +614,7 @@ info(int all, int longout, int group, int label)
             printf("Empty (tag %d) : %d slots.\n", DFTAG_NULL, empty);
     }
     else {
-        mark = (int *)calloc(he_numDesc, sizeof(int));
+        mark = (int *)calloc((size_t)he_numDesc, sizeof(int));
 
         if (all) {
             start = 0;
@@ -823,12 +823,12 @@ struct {
 HE_FUNC
 findFunc(char *fword)
 {
-    unsigned len;
-    int      found = -1;
+    size_t len;
+    int    found = -1;
 
     len = strlen((const char *)fword);
 
-    for (int i = 0; i < sizeof(he_funcTab) / sizeof(he_funcTab[0]); i++)
+    for (size_t i = 0; i < sizeof(he_funcTab) / sizeof(he_funcTab[0]); i++)
         if (!strncmp(he_funcTab[i].str, (const char *)fword, len)) {
             /* check for exact match */
             if (strlen(he_funcTab[i].str) == len)
@@ -1225,12 +1225,12 @@ struct {
 int
 findKey(char *word)
 {
-    unsigned len;
-    int      found = -1;
+    size_t len;
+    int    found = -1;
 
     len = strlen(word);
 
-    for (int i = 0; i < sizeof(he_keyTab) / sizeof(he_keyTab[0]); i++)
+    for (size_t i = 0; i < sizeof(he_keyTab) / sizeof(he_keyTab[0]); i++)
         if (!strncmp(he_keyTab[i].str, word, len)) {
             /* if this is an exact match, just return */
             if (strlen(he_keyTab[i].str) == len)
@@ -1408,7 +1408,7 @@ nextToken(char **p)
         while (*s && !isalnum((int)*s))
             s++;
 
-    q = tok = (char *)malloc((s - (*p)) + 1);
+    q = tok = (char *)malloc((size_t)(s - (*p)) + 1);
     while (*p != s)
         *q++ = *(*p)++;
     *q = '\0';
